prevent passing invalid cwd to Popen
Bizarrely, this manifests as a failure to find normal commands like bash,
looks like a bug in Popen when it is passed invalid cwd=.

Worryingly, cwd may become invalid after this check.
So there is a late assert to pinpoint the problem rather than just
failing to find bash. Even then cwd may become invalid after that, but
I can't fix it unless Popen is fixed either not to butcher PATH when it gets
a bad cwd, or otherwise to raise a more distinctive error.

Because I'm not about to reinvent ``which`` in Python just to hand-feed
Popen an absolute path to the executable, that sucks.