complain properly about abs path as venv name
This is a continuation of the issue with giving absolute pathnames
as the positional name of a virtualenv, NOT using --path.
It happens because os.path.join silently discards earlier
path segments when a later one was absolute.

There was an assert checking if ve_name == ve_path,
but users should never see this.

The assert was also erroneously catching the case where
you are sitting in $WORKON_HOME (e.g.: cd ~/.virtualenvs)