fix up functional tests relying on blank $HOME
Can't use PatchedModule to ensure that expanduser('~')
returns '' (and thus we default past the real HOME)
since we are out-of-process.
So using a spurious $HOME allows the real $HOME path to be avoided
instead of leaking into the test
(e.g. test failure because it looks at actual /home/sasha/.vexrc
instead of defaulting to the test stuff)