be paranoid about ve_base being a real path
The worst case here is that someone puts arbitrary shell junk
$WORKON_HOME, which is then propagated into ve_base, and
then replaced into the output of vex --shell-config, which output
is then evaluated by user's shell, doing whatever was in the shell
junk.

Instead of endlessly blacklisting characters which might indicate danger
(and missing malicious stuff that just avoids them)
we just make sure it's a directory path. This isn't a complete solution insofar
as it is possible to craft paths which could be evaluated stupidly,
this is also a concern up to the actual shell config snippets

in tests: patch out existence checks in tests of get_ve_base.
this allows get_ve_base to check for existence if it
wants to, without the tests failing because the paths
are actually fictitious