don't sub $WORKON_HOME with scary-looking paths
This is intended as defense-in-depth behind checks that the value of
WORKON_HOME is a real path; for example, it is remotely feasible
that a real path could be crafted which also does something stupid
when evaluated as part of the emitted shell snippet.

The whole issue is defense-in-depth behind whatever means
you are normally using to prevent things like 'EDITOR=rm',
if those fail you are in trouble anyway even without vex.

includes tests for the function which decides if strings look scary as paths;
these aren't exhaustive, they just provide a baseline.

scary_path explicitly works on bytestrings